Release 3.0.0 (#21)
* intial PlatformIO configuration (#1)

* initial TCP integration using Arduino WiFi module (#2)

* adding Command Prompt functionality for interpreting commands from GUI (#3)

* adding classes for the Hall Effect sensors and functionality to test data collection (#4)

* creating initial class to abstract  WiFi and Server functionality (#6)

* updating begin/end byte to align with GUI (#7)

* Code Refactoring (#8)

* updating preprocessor directives

* moving header files for better organization

* reduced overhead of ESPServer class and passed values by reference

* creating hidden secrets header for wifi credentials

* initial L9110H Motor driver (#9)

* initial ESP Server modularization (#10)

* Feature server composition (#11)

* initial ESP Server modularization

* resolving format of data when sending to client

* resolving format of data when sending to client

* removing hidden merge conflict messages by VS Code

* Feature: Threaded Functionality (#13)

* moved tactile data reading functionality to main loop to prevent blocking incoming commands

* removed PWM functionality for test motor to align with gripper motor

* condensed GUI and Serial command processing and removed unused functionality

* adding method to resize buffer message to GUI request (#14)

* Update ReadMe (#15)

* adding initial Overview and Table of Contents

* adding section for Installing and Running the application

* adding architecture diagram and datasheet for tactile sensor

* adding Repository Folder Structure

* adding Helpful Articles

* fixing newlines in Repository Folder Structure

* adding header to architecture diagram

* updating table of content links and adding project background section

* updating Architecture content and diagram

* fixing typos in architecture

* updating context in sensors

* fixing typo in readme

* fixing typos

* GUI sensor calibration (#17)

* adding calibrate method

* adding sensor calibration

* Feature - collection method (#18)

* decoupling stopping gripper and motor delay with open and close commands

* expanding process_command functionality and including new collection method

* adding persistent gui client parameter to new send method

* updating class descriptions

* Refactoring of Tactile Sensor collection methods (#19)

* refactoring read and calibrate methods for MLX90393 sensor and adding calibration on initialization

* abstracting read data method and adding averaging method to TactileSensor

* refactoring read data methods for better abstraction

* refactoring Tactile Sensor methods and adding method to read max z values

* updating send data method to use readDataMaxZ

* cleanup of unused variables and functions

* Finite State Machine (#20)

* work in progress: stashing changes

* stashing changes

* updating FSM design pattern

* refactoring state machine design pattern to FSM and removing old ESP server classes

* updating details in readme

+++ include/vector3.h
@@ -1,11 +1,11 @@
 #ifndef vector3_h
 #define vector3_h
 
-struct vector3{
-  int x, y, z;
+struct vector3 {
+  float x, y, z;
 };
 
-struct vector3Double{
+struct vector3Double {
   double x, y, z;
 };
